feat(OpenChemLibEditor): Add Atom Selection Listener with Reaction Support (#1)
* feat(OpenChemLibEditor): implemented atom selection listener

Implemented an AtomSelectionEvent which is fired when the atom selection state changes. This event exposes the selected atoms via an ArrayList of selected atom indices and can be interacted with using an atom selection listener in the OpenChemLibEditor. The AtomSelectionEvent supports selected atoms in reactions & molecules. In REACTION mode, selected atoms are returned as a map with two keys, 'reactants' and 'products'. The values of those are a list of list where the outer list contains molecules and the inner list contains selected atoms.