feat: Implement Bond Highlighting and Atom Coloring (#2)
* feat(OpenChemLibEditor): implemented atom selection listener

Implemented an AtomSelectionEvent which is fired when the atom selection state changes. This event exposes the selected atoms via an ArrayList of selected atom indices and can be interacted with using an atom selection listener in the OpenChemLibEditor.

* feat(OpenChemLibEditor): extended atom selection listener

Extended the AtomSelectionEvent to support selected atoms in reactions. In REACTION mode, selected atoms are returned as a map with two keys, 'reactants' and 'products'. The values of those are a list of list where the outer list contains molecules and the inner list contains selected atoms.

* feat: implement bond highlighting and atom coloring

Implemented programmatic bond background and foreground highlighting and coloring of atoms. For this, the following methods were exposed in the OpenChemLibEditor:
- highlightBondsBackground: Highlights bonds by painting a blue bar behind them.
- highlightBondsForeground: Highlights bonds by painting them orange.
- clearHighlights: Removes all highlighting in a molecule.
- setAtomColor: Sets the color for a specific atom in a molecule.
- removeAtomColors: Removes all custom coloring from a molecule.

Note: These features were implemented to work with the StructureEditor and the ReactionEditor. However, due to current behavior of openchemlib-js, they do NOT yet work for the ReactionEditor.