Enhance Atom Property Customization and Labeling in OpenChemLibEditor (#3)
* feat: implement custom atom labeling

This commit exposed the 'setAtomCustomLabel' method to the OpenChemLibEditor. Currently, only editors in 'MOLECULE' mode are supported.

* feat: implement automatic setting of atom properties

Implemented the automatic transfer of custom atom properties from Java to JS. This means that if a StereoMolecule which is passed to 'setValue()' of the StructureEditor contains custom atom labels, atom colors, or bond highlights, they will automatically be passed on to the JS editor and will get displayed accordingly.